+++ Packages/com.quickeye.ui-toolkit-plus/Editor/UxmlClassGeneration/PostHeaderDrawer.cs
@@ -1,0 +1,14 @@
+namespace QuickEye.UIToolkit.Editor
+{
+    public abstract class PostHeaderDrawer
+    {
+        public UnityEditor.Editor Editor { get; private set; }
+
+        public PostHeaderDrawer(UnityEditor.Editor editor)
+        {
+            Editor = editor;
+        }
+        
+        public abstract void OnGUI();
+    }
+}